% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_defactorize.R
\name{df_defactorize}
\alias{df_defactorize}
\title{df_defactorize}
\usage{
df_defactorize(df)
}
\arguments{
\item{df}{a data.frame like object}
}
\value{
returns the same data.frame except that factor columns have been transformed into 
        character columns
}
\description{
df_defactorize
}
\examples{

df <- 
  data.frame(
    a = 1:2, 
    b = factor(c("a", "b")), 
    c = as.character(letters[3:4]), 
    stringsAsFactors = FALSE
 )
vapply(df, class, "")

df_df <- df_defactorize(df)
vapply(df_df, class, "")

}
