% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mode_multi.R
\name{stats_mode_multi}
\alias{stats_mode_multi}
\title{Mode Allowing for Multi Modal Mode}
\usage{
stats_mode_multi(x)
}
\arguments{
\item{x}{vector to get mode for}
}
\value{
vector with all modes
}
\description{
Function calculating the mode, allowing for multiple modes in case of equal frequencies.
}
