% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_modus.R
\name{stats_modus}
\alias{stats_modus}
\title{stats_modus}
\usage{
stats_modus(x, multimodal = FALSE, warn = TRUE)
}
\arguments{
\item{x}{vector to get mode for}

\item{multimodal}{wether or not all modes should be returned in case of more than one}

\item{warn}{should the function warn about multimodal outcomes?}
}
\description{
Function calculating the mode.
}
