% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_modus_multi.R
\name{stats_modus_multi}
\alias{stats_modus_multi}
\title{stats_modus_multi}
\usage{
stats_modus_multi(x)
}
\arguments{
\item{x}{vector to get mode for}
}
\value{
return vector with
}
\description{
stats_modus_multi
}
