% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stamp.R
\name{time_stamp}
\alias{time_stamp}
\title{Time Stamps for File Names}
\usage{
time_stamp(ts = Sys.time(), sep = c("-", "_", "_"))
}
\arguments{
\item{ts}{one or more POSIX time stamp}

\item{sep}{separators to be used for formatting}
}
\value{
Returns timestamp string in format yyyy-mm-dd_HH_MM_SS ready to be used 
safely in file names on various operating systems.
}
\description{
Generating file name ready iso time stamps.
}
\examples{

time_stamp()
time_stamp( Sys.time() - 10000 )


}
