% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_i_fit_index.R
\name{tool_i_fit_index}
\alias{tool_i_fit_index}
\title{Fit number into length of index}
\usage{
tool_i_fit_index(i, index)
}
\arguments{
\item{i}{number to make sure to fit into index range}

\item{index}{size of index}
}
\value{
number that fits into index by circularly mapping i to range of 1 to size of index
}
\description{
Fit number into length of index
}
\examples{

tool_i_fit_index(-2:6, 3)

}
\seealso{
\link{tool_i_fit_obj}
}
