% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_i_fit_obj.R
\name{tool_i_fit_obj}
\alias{tool_i_fit_obj}
\title{Subset object even if index is out of range}
\usage{
tool_i_fit_obj(i, obj)
}
\arguments{
\item{i}{number to make sure to fit into index range}

\item{obj}{object to subset data from}
}
\value{
elements of object circularly indexed by  mapping i to range of 1 to size of object
}
\description{
Subset object even if index is out of range
}
\examples{

tool_i_fit_obj(-2:6, 3)

}
\seealso{
\link{tool_i_fit_index}
}
