% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_gen_param_list.R
\name{web_gen_param_list_expand}
\alias{web_gen_param_list_expand}
\title{URL Parameter Combinations}
\usage{
web_gen_param_list_expand(..., sep_1 = "=", sep_2 = "&")
}
\arguments{
\item{...}{multiple vectors passed on as named arguments or a single list or a data.frame}

\item{sep_1}{first separator to use between key and value}

\item{sep_2}{second separator to use between key-value pairs}
}
\value{
string vector with assembled query string parameter combinations
}
\description{
Generate URL parameter combinations from sets of parameter values.
}
\examples{

web_gen_param_list_expand(q = "beluga", lang = c("de", "en"))

}
