% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stamp.R
\name{time_stamp}
\alias{time_stamp}
\title{time_stamp}
\usage{
time_stamp(ts = Sys.time(), sep = c("-", "_", "_"))
}
\arguments{
\item{ts}{one or many POSIX time stamp}

\item{sep}{separators to be used for formatting}
}
\description{
Generating file name ready iso time stamps.
}
\examples{

time_stamp()
time_stamp( Sys.time() - 10000 )

}
