% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stamp.R
\name{time_stamp}
\alias{time_stamp}
\title{time_stamp}
\usage{
time_stamp(ts = Sys.time())
}
\arguments{
\item{ts}{one or many POSIX timestamp}
}
\description{
time_stamp
}
\examples{

time_stamp()
time_stamp( Sys.time() - 10000 )

}
