% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_gen_param_list.R
\name{web_gen_param_list_exapnd}
\alias{web_gen_param_list_exapnd}
\title{web_gen_param_list_exapnd}
\usage{
web_gen_param_list_exapnd(..., sep_1 = "=", sep_2 = "&")
}
\arguments{
\item{...}{}

\item{sep_1}{se}

\item{sep_2}{}
}
\description{
Generate an url parameter combination from sets of parameter values.
}
\examples{

web_gen_param_list_exapnd(q = "beluga", lang = c("de", "en"))

}
